move factories into tests dir
